Make serf capable of checking out from Subversion 1.3 repositories with the/a "lost response" with lingering close bug. (See also r858579)
This code is not pretty but it is a safe and local fix to this problem.
Fixing this problem in a clean way probably requires some design changes to
the request creation in ra_serf.

Before fixing the calculation bug in r1154733 checking out neon from its
repository was completely impossible. But after that fix checkout still failed
in about 30-50% of the cases. This patch fixes those remaining failures by
applying the same fix to the xml requests.

* subversion/libsvn_ra_serf/ra_serf.h
  (svn_ra_serf__xml_parser_t): Add 3 variables to allow tracking restarted
    connections.

* subversion/libsvn_ra_serf/util.c
  (svn_ra_serf__handle_xml_parser): Detect restarted requests by checking if
    the response headers changed since the last invocation. If so apply the same
    skip trick as used in ra_serf/update.c to avoid reading the same xml twice,
    which makes the xml invalid as <?xml is only valid at the start of the xml.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1155160 13f79535-47bb-0310-9956-ffa450edef68