Remove the SQLITE_OPEN_SHAREDCACHE flag from opening SQLite databases. VisualSVN and AnkhSVN users are reporting errors caused by the limitations of the shared caching in SQLite.
The current implementation temporarily doesn't allow certain operations that
are allowed concurrently if the usual via-file sharing is used. Until this is
fixed in SQLite I think we should stop using this flag.
The common errors are 'database table is locked' and 'schema has changed'.

VisualSVN already disabled this flag in their build (reported by Ivan) and I
will have to do the same thing in SharpSvn and as far as I know this are
currently the two most prominent multithreaded subversion library users
that would benefit if this flag always worked correctly.

I enabled this flag during the development of Subversion 1.7 to speed up
the then common case of opening the same wc.db file multiple times.
Currently we don't open the same database multiple times from a single
codepath. Instead we share the connection handle properly.

* subversion/libsvn_subr/sqlite.c
  (internal_open): Remove SQLITE_OPEN_SHAREDCACHE flag as noted above.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1185242 13f79535-47bb-0310-9956-ffa450edef68