Remove editor-test.
There was only one test in that file, and it turns out our testing strategy
of inserting shims is much easier and gives us better coverage.  Rather than
try to keep this test up with a moving target, I'm removing it now and we
can bring it back if it is later needed.

* build.conf
  (editor-test): Remove.
  (libs): Remove editor-test.

* subversion/tests/libsvn_delta/editor-test.c:
  Remove.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1203343 13f79535-47bb-0310-9956-ffa450edef68