Finish the server-side bit of issue #4082.
Check for a flag in the client's XML REPORT request for log revisions
which means "encode any binary property values before trying to send
them across the wire", and honor that request.  We'll use base64
encoding, just as we do in other portions of mod_dav_svn.  Clients
which don't request this encoding of binary properties will have to
endure the previous behavior for compatibility's sake.

* subversion/mod_dav_svn/reports/log.c
  (struct log_receiver_baton): Add 'encode_binary_props' member.
  (log_receiver): If the client supports binary property encoding, and
    we've a non-XML-safe value to transmit, then base64-encode that
    sucker when embedding it into our XML response.
  (dav_svn__log_report): Initialize the baton's 'encode_binary_props'
    member, and check for the XML tag which means "Please encode
    binary property values."


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1215260 13f79535-47bb-0310-9956-ffa450edef68