When comparing output from svnsync tests, use the same binary which generated the dump output to generate the dump from the target repo.
Because the dump format is not unique (nodes may appear in a somewhat arbitrary
order), dump files produced by prior versions of Subversion, while still
correct, may not compare correctly when compared line-wise.  In some cases,
this adds a bit of a performance penalty to the tests, but is more correct.

* subversion/tests/cmdline/svnsync_tests.py
  (verify_mirror): Verify to sandboxes, rather than a sandbox against expected
    dumpfile contents.
  (run_test): Use the two existing sandboxes for comparison, or create a new
    one if a specific dumpfile is required.
  (delete_revprops): Create a sandbox to compare against.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1230676 13f79535-47bb-0310-9956-ffa450edef68