Ev2 shims: Truthfully report our base checksum as being that of the empty stream.
Note: This breaks several assumptions in various delta-editor receivers about
the validity of this checksum.  These have been patched to ignore the checksum
if it is that of the empty stream.  This will not affect correctness, as the
final checksum, as supplied by close_file() is still used to detect corruption,
and it hasn't changed.

Current number of Ev2 test failures: 11
(This fixes svnrdump test 35, but causes svnrdump test 38 to fail.)

[Note from the future: This was reverted in r1241531, which see for a reference
 to the detailed discussion.]

* subversion/libsvn_fs_fs/tree.c
  (apply_textdelta): Ignore the empty checksum.

* subversion/libsvn_fs_base/tree.c
  (txn_body_apply_textdelta): Same.

* subversion/libsvn_wc/update_editor.c
  (apply_textdelta): Same.

* subversion/libsvn_wc/externals.c
  (apply_textdelta): Same.

* subversion/libsvn_client/repos_diff.c
  (apply_textdelta): Same.

* subversion/libsvn_delta/compat.c
  (drive_tree): Supply the checksum of the empty stream to apply_textdelta(),
    as we are always delta'ing against the empty stream by this point.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1241097 13f79535-47bb-0310-9956-ffa450edef68